use less strings.suffix and move okexapi/type.go to package okex
